Move ECL utilities to separate file.
