+++ opm/utility/ECLGraph.cpp
@@ -23,6 +23,7 @@
 #endif
 
 #include <opm/utility/ECLGraph.hpp>
+#include <opm/utility/ECLUtilities.hpp>
 
 #include <algorithm>
 #include <array>
@@ -47,631 +48,6 @@
 ///
 /// Implementation of \c ECLGraph interface.
 
-namespace {
-    namespace ECL {
-        using GridPtr = ::ERT::ert_unique_ptr<ecl_grid_type, ecl_grid_free>;
-        using FilePtr = ::ERT::ert_unique_ptr<ecl_file_type, ecl_file_close>;
-
-        /// Retrieve global pore-volume vector from INIT source.
-        ///
-        /// Specialised tool needed to determine the active cells.
-        ///
-        /// \param[in] G ERT Grid representation.
-        ///
-        /// \param[in] init ERT representation of INIT source.
-        ///
-        /// \return Vector of pore-volumes for all global cells of \p G.
-        std::vector<double>
-        getPVolVector(const ecl_grid_type* G,
-                      const ecl_file_type* init);
-
-        /// Internalise on-disk representation of ECLIPSE grid.
-        ///
-        /// \param[in] grid Name or prefix of on-disk representation of
-        ///                 ECLIPSE grid.  If using a prefix, the loader
-        ///                 will consider both .EGRID and .GRID versions of
-        ///                 the input.
-        ///
-        /// \return Internalised ERT Grid representation.
-        GridPtr loadCase(const boost::filesystem::path& grid);
-
-        /// Internalise on-disk representation of ECL file.
-        ///
-        /// \param[in] file Name of ECLIPSE output file.
-        ///
-        /// \return Internalised ERT file contents.
-        FilePtr loadFile(const boost::filesystem::path& file);
-
-        /// Extract non-neighbouring connections from ECLIPSE model
-        ///
-        /// \param[in] G ERT Grid representation.
-        ///
-        /// \param[in] init ERT representation of INIT source.
-        ///
-        /// \return Model's non-neighbouring connections, including those
-        ///         between main and local grids.
-        std::vector<ecl_nnc_type>
-        loadNNC(const ecl_grid_type* G,
-                const ecl_file_type* init);
-
-        /// Facility for mapping a result set vector defined either on all
-        /// global cells or on the explicitly active cells (ACTNUM != 0) to
-        /// the global cells.
-        class ScatterMap
-        {
-        public:
-            /// Constructor
-            ///
-            /// \param[in] G ERT Grid representation.
-            ///
-            /// \param[in] pvol Vector of pore-volumes on all global cells
-            ///                 of \p G.  Typically obtained through
-            ///                 function getPVolVector().
-            ScatterMap(const ecl_grid_type*       G,
-                       const std::vector<double>& pvol);
-
-            /// Retrive global cell indices of all active cells.
-            std::vector<std::size_t> activeGlobalCells() const;
-
-            /// Map input vector to all global cells.
-            ///
-            /// \param[in] x Input vector, defined on the explicitly active
-            ///              cells, all global cells or some other subset
-            ///              (e.g., all non-neighbouring connections).
-            ///
-            /// \return Input vector mapped to global cells or unchanged if
-            /// input is defined on some other subset.
-            template <typename T>
-            std::vector<T>
-            scatterToGlobal(const std::vector<T>& x) const;
-
-        private:
-            /// Explicit mapping between ACTNUM!=0 cells and global cells.
-            struct ID {
-                std::size_t act;
-                std::size_t glob;
-            };
-
-            /// Number of explicitly active cells (SUM(ACTNUM != 0)).
-            std::size_t nact_;
-
-            /// Number of global cells.
-            std::size_t nglob_;
-
-            /// Active subset of global cells.
-            std::vector<ID> subset_;
-        };
-
-        /// Facility for cached querying of ECLIPSE result set.
-        class GlobalCellData
-        {
-        public:
-            /// Constructor
-            ///
-            /// \param[in] G ERT Grid representation.
-            ///
-            /// \param[in] pvol Vector of pore-volumes on all global cells
-            ///                 of \p G.  Typically obtained through
-            ///                 function getPVolVector().
-            GlobalCellData(const ecl_grid_type*       grid,
-                           const std::vector<double>& pvol);
-
-            /// Assign result set backing object.
-            ///
-            /// \param[in] src Name of result set (restart) file, possibly
-            /// unified.
-            void assignDataSource(const boost::filesystem::path& src);
-
-            /// Assign result set backing object.
-            ///
-            /// Specialised convenience interface.
-            ///
-            /// \param[in] src Internalised result set.  Typically
-            /// corresponds to the INIT object.
-            void assignDataSource(FilePtr src);
-
-            using DataVector = std::vector<double>;
-
-            /// Retrive ERT representation of current result set.
-            ///
-            /// Convenience method that typically accesses the result set
-            /// that was assigned through assignDataSource() with a \c
-            /// FilePtr argument.
-            const ecl_file_type* getDataSource() const;
-
-            /// Retrive global cell indices of all active cells.
-            std::vector<std::size_t> activeGlobalCells() const;
-
-            /// Predicate for whether or not a particular vector exists in
-            /// the current result set.
-            bool haveVector(const std::string& vector) const;
-
-            /// Retrieve particular vector from current result set.
-            ///
-            /// \param[in] vector Name of result vector.  Clients should
-            ///                   ensure existence of data through
-            ///                   haveVector() before calling getVector().
-            ///
-            /// \param[in] occurrence Selected temporal vector.  Essentially
-            ///                       the report step number.
-            ///
-            /// \return Result vector corresponding to named quantity at
-            /// given time.
-            const DataVector&
-            getVector(const std::string& vector,
-                      const int          occurrence = 0);
-
-        private:
-            using DataHandle     = std::unique_ptr<DataVector>;
-            using TemporalData   = std::map<int, DataHandle>;
-            using DataCollection = std::map<std::string, TemporalData>;
-
-            /// Map results from active to global cells.
-            ScatterMap map_;
-
-            /// Current result set.
-            ECL::FilePtr src_;
-
-            /// Result vector cache.
-            DataCollection coll_;
-
-            /// Load a result vector from backing store.
-            ///
-            /// \param[in] vector Name of result vector.  Clients should
-            ///                   ensure existence of data through
-            ///                   haveVector() before calling getVector().
-            ///
-            /// \param[in] occurrence Selected temporal vector.  Essentially
-            ///                       the report step number.
-            ///
-            /// \return Raw result vector as represented in the backing
-            /// store.  Must be mapped to global cells prior to access.
-            DataVector
-            load(const std::string& vector,
-                 const int          occurrence);
-        };
-    } // namespace ECL
-
-    /// Derive neighbourship relations between active cells from Cartesian
-    /// relations.
-    class CellCollection
-    {
-    public:
-        /// Canonical directions of Cartesian neighbours.
-        enum class Direction { I, J, K };
-
-        /// Constructor.
-        ///
-        /// \param[in] G ERT Grid representation.
-        ///
-        /// \param[in] active_glob_cells Global indices of model's active
-        ///       cells (\code ACTNUM != 0 && pore_volume > 0 \endcode).
-        CellCollection(const ecl_grid_type*       G,
-                       std::vector<std::size_t>&& active_glob_cells);
-
-        /// Retrieve global indices of model's active cells.
-        ///
-        /// Convenience method that returns the second constructor argument.
-        const std::vector<std::size_t>& activeGlobalCells() const;
-
-        /// Retrieve total number of cells in model, including inactive ones.
-        ///
-        /// Needed to allocate result vectors on global cells.
-        std::size_t numGlobalCells() const;
-
-        /// Retrieve active cell ID of particular global cell.
-        ///
-        /// \param[in] globalCell Index of particular global cell.
-        ///
-        /// \return Active cell ID of \p globalCell.  Returns negative one
-        /// (\code -1 \endcode) if \code globalCell >= numGlobalCells
-        /// \endcode or if the global cell is inactive.
-        int activeCell(const std::size_t globalCell) const;
-
-        /// Retrieve active cell ID of particular global cell's neighbour in
-        /// given Cartesian direction.
-        ///
-        /// \param[in] globalCell Index of particular global cell.
-        ///
-        /// \param[in] d Cartesian direction in which to look for a
-        /// neighbouring cell.
-        ///
-        /// \return Active cell ID of \p globalCell's neighbour in direction
-        /// \d.  Returns negative one (\code -1 \endcode) if \code
-        /// globalCell >= numGlobalCells \endcode or if the global cell is
-        /// inactive, or if there is no neighbour in direction \p d (e.g.,
-        /// if purported neighbour would be outside model).
-        int cartesianNeighbour(const std::size_t globalCell,
-                               const Direction   d) const;
-
-    private:
-        using IJKTuple = std::array<std::size_t,3>;
-
-        /// Number of global cells in X (I) direction.
-        std::size_t nx_;
-
-        /// Number of global cells in Y (J) direction.
-        std::size_t ny_;
-
-        /// Number of global cells in Z (K) direction.
-        std::size_t nz_;
-
-        /// Global indices of model's active cells.
-        std::vector<std::size_t> glob_cell_;
-
-        /// Active cell ID of model's global cells.  Negative one (\code -1
-        /// \endcode) if inactive.
-        std::vector<int> active_id_;
-
-        /// Retrieve number of active cells.  Equivalent to \code
-        /// glob_cell_.size() \endcode, but with a named expression.
-        std::size_t numActive() const;
-
-        /// Compute linear index of global cell from explicit (I,J,K) tuple.
-        ///
-        /// \param[in] ijk Explicit (I,J,K) tuple of global cell.
-        ///
-        /// \return Linear index (natural ordering) of global cell (I,J,K).
-        std::size_t globIdx(const IJKTuple& ijk) const;
-
-        /// Decompose global (linear) cell index into its (I,J,K) index
-        /// tuple.
-        ///
-        /// \param[in] globalCell Index of particular global cell.  Must be
-        ///    in the range \code [0 .. numGlobalCells()) \endcode.
-        ///
-        /// \return Index triplet of \p globalCell's location within model.
-        IJKTuple ind2sub(const std::size_t globalCell) const;
-    };
-} // Anonymous namespace
-
-// ======================================================================
-
-std::vector<double>
-ECL::getPVolVector(const ecl_grid_type* G,
-                   const ecl_file_type* init)
-{
-    auto make_szt = [](const int i)
-    {
-        return static_cast<std::vector<double>::size_type>(i);
-    };
-
-    const auto nx = make_szt(ecl_grid_get_nx(G));
-    const auto ny = make_szt(ecl_grid_get_ny(G));
-    const auto nz = make_szt(ecl_grid_get_nz(G));
-
-    const auto nglob = nx * ny * nz;
-
-    auto pvol = std::vector<double>(nglob, 1.0);
-
-    if (ecl_file_has_kw(init, "PORV")) {
-        auto porv =
-            ecl_file_iget_named_kw(init, "PORV", 0);
-
-        assert ((make_szt(ecl_kw_get_size(porv)) == nglob)
-                && "Pore-volume must be provided for all global cells");
-
-        ecl_kw_get_data_as_double(porv, pvol.data());
-    }
-
-    return pvol;
-}
-
-ECL::GridPtr
-ECL::loadCase(const boost::filesystem::path& grid)
-{
-    auto G = GridPtr{
-        ecl_grid_load_case(grid.generic_string().c_str())
-    };
-
-    if (! G) {
-        std::ostringstream os;
-
-        os << "Failed to load ECL Grid from '"
-           << grid.generic_string() << '\'';
-
-        throw std::invalid_argument(os.str());
-    }
-
-    return G;
-}
-
-ECL::FilePtr
-ECL::loadFile(const boost::filesystem::path& file)
-{
-    // Read-only, keep open between requests
-    const auto open_flags = 0;
-
-    auto F = FilePtr{
-        ecl_file_open(file.generic_string().c_str(), open_flags)
-    };
-
-    if (! F) {
-        std::ostringstream os;
-
-        os << "Failed to load ECL File object from '"
-           << file.generic_string() << '\'';
-
-        throw std::invalid_argument(os.str());
-    }
-
-    return F;
-}
-
-std::vector<ecl_nnc_type>
-ECL::loadNNC(const ecl_grid_type* G,
-             const ecl_file_type* init)
-{
-    auto make_szt = [](const int n)
-    {
-        return static_cast<std::vector<ecl_nnc_type>::size_type>(n);
-    };
-
-    auto nncData = std::vector<ecl_nnc_type>{};
-
-    const auto numNNC = make_szt(ecl_nnc_export_get_size(G));
-
-    if (numNNC > 0) {
-        nncData.resize(numNNC);
-
-        ecl_nnc_export(G, init, nncData.data());
-    }
-
-    std::sort(nncData.begin(), nncData.end(),
-              [](const ecl_nnc_type& nd1, const ecl_nnc_type& nd2) {
-                  return nd1.input_index < nd2.input_index;
-              });
-
-    return nncData;
-}
-
-// ======================================================================
-
-ECL::ScatterMap::ScatterMap(const ecl_grid_type*       G,
-                            const std::vector<double>& pvol)
-{
-    auto make_szt = [](const int i)
-    {
-        return static_cast<std::size_t>(i);
-    };
-
-    this->nact_ = make_szt(ecl_grid_get_nactive(G));
-
-    {
-        const auto nx = make_szt(ecl_grid_get_nx(G));
-        const auto ny = make_szt(ecl_grid_get_ny(G));
-        const auto nz = make_szt(ecl_grid_get_nz(G));
-
-        this->nglob_ = nx * ny * nz;
-    }
-
-    subset_.clear();
-    subset_.reserve(this->nact_);
-
-    if (pvol.empty()) {
-        for (decltype(ecl_grid_get_nactive(G))
-                 act = 0, nact = ecl_grid_get_nactive(G);
-             act < nact; ++act)
-        {
-            const auto glob =
-                make_szt(ecl_grid_get_global_index1A(G, act));
-
-            this->subset_.push_back(ID{ make_szt(act), glob });
-        }
-    }
-    else {
-        for (decltype(ecl_grid_get_nactive(G))
-                 act = 0, nact = ecl_grid_get_nactive(G);
-             act < nact; ++act)
-        {
-            const auto glob =
-                make_szt(ecl_grid_get_global_index1A(G, act));
-
-            if (pvol[glob] > 0.0) {
-                this->subset_.push_back(ID{ make_szt(act), glob });
-            }
-        }
-    }
-}
-
-std::vector<std::size_t>
-ECL::ScatterMap::activeGlobalCells() const
-{
-    auto active = std::vector<std::size_t>{};
-    active.reserve(this->subset_.size());
-
-    for (const auto& id : this->subset_) {
-        active.push_back(id.glob);
-    }
-
-    return active;
-}
-
-template <typename T>
-std::vector<T>
-ECL::ScatterMap::scatterToGlobal(const std::vector<T>& x) const
-{
-    // Assume that input vector 'x' is either defined on explicit
-    // notion of active cells (ACTNUM != 0) or on all global cells or
-    // some other contiguous index set (e.g., the NNCs).
-
-    if (x.size() != static_cast<decltype(x.size())>(this->nact_)) {
-        // Input not defined on explictly active cells.  Let caller
-        // deal with it.  This typically corresponds to the set of
-        // global cells or the list of NNCs.
-        return x;
-    }
-
-    auto y = std::vector<T>(this->nglob_);
-
-    for (const auto& i : this->subset_) {
-        y[i.glob] = x[i.act];
-    }
-
-    return y;
-}
-
-// ======================================================================
-
-ECL::GlobalCellData::GlobalCellData(const ecl_grid_type*       grid,
-                                    const std::vector<double>& pvol)
-    : map_(grid, pvol)
-{
-}
-
-void
-ECL::GlobalCellData::assignDataSource(const boost::filesystem::path& src)
-{
-    this->src_ = ECL::loadFile(src);
-}
-
-void
-ECL::GlobalCellData::assignDataSource(ECL::FilePtr src)
-{
-    this->src_ = std::move(src);
-}
-
-const ecl_file_type*
-ECL::GlobalCellData::getDataSource() const
-{
-    return this->src_.get();
-}
-
-std::vector<std::size_t>
-ECL::GlobalCellData::activeGlobalCells() const
-{
-    return this->map_.activeGlobalCells();
-}
-
-bool
-ECL::GlobalCellData::haveVector(const std::string& vector) const
-{
-    return ecl_file_has_kw(this->src_.get(), vector.c_str());
-}
-
-const ECL::GlobalCellData::DataVector&
-ECL::GlobalCellData::getVector(const std::string& vector,
-                               const int          occurrence)
-{
-    auto& p = this->coll_[vector][occurrence];
-
-    if (! p) {
-        auto x = this->map_.scatterToGlobal(load(vector, occurrence));
-
-        p = DataHandle{ new DataVector{ std::move(x) } };
-    }
-
-    return *p;
-}
-
-ECL::GlobalCellData::DataVector
-ECL::GlobalCellData::load(const std::string& vector,
-                          const int          occurrence)
-{
-    auto x = ecl_file_iget_named_kw(this->src_.get(),
-                                    vector.c_str(),
-                                    occurrence);
-
-    auto y = DataVector(ecl_kw_get_size(x));
-
-    ecl_kw_get_data_as_double(x, y.data());
-
-    return y;
-}
-
-// ======================================================================
-
-CellCollection::CellCollection(const ecl_grid_type*       G,
-                               std::vector<std::size_t>&& active_glob_cells)
-    : nx_       (static_cast<std::size_t>(ecl_grid_get_nx(G)))
-    , ny_       (static_cast<std::size_t>(ecl_grid_get_ny(G)))
-    , nz_       (static_cast<std::size_t>(ecl_grid_get_nz(G)))
-    , glob_cell_(std::move(active_glob_cells))
-    , active_id_(nx_ * ny_ * nz_, -1)
-{
-    auto act = 0;
-
-    for (const auto& id : glob_cell_) {
-        this->active_id_[static_cast<std::size_t>(id)] = act++;
-    }
-}
-
-const std::vector<std::size_t>&
-CellCollection::activeGlobalCells() const
-{
-    return this->glob_cell_;
-}
-
-std::size_t
-CellCollection::numActive() const
-{
-    return this->glob_cell_.size();
-}
-
-std::size_t
-CellCollection::numGlobalCells() const
-{
-    return this->active_id_.size();
-}
-
-int
-CellCollection::activeCell(const std::size_t globalCell) const
-{
-    if (globalCell >= numGlobalCells()) { return -1; }
-
-    return this->active_id_[globalCell];
-}
-
-int
-CellCollection::cartesianNeighbour(const std::size_t globalCell,
-                                   const Direction   d) const
-{
-    if (globalCell >= numGlobalCells()) { return -1; }
-
-    auto ijk = ind2sub(globalCell);
-
-    if      (d == Direction::I) { ijk[0] += 1; }
-    else if (d == Direction::J) { ijk[1] += 1; }
-    else if (d == Direction::K) { ijk[2] += 1; }
-    else {
-        return -1;
-    }
-
-    const auto globNeigh = globIdx(ijk);
-
-    if (globNeigh >= numGlobalCells()) { return -1; }
-
-    return this->active_id_[globNeigh];
-}
-
-std::size_t
-CellCollection::globIdx(const IJKTuple& ijk) const
-{
-    if (ijk[0] >= this->nx_) { return -1; }
-    if (ijk[1] >= this->ny_) { return -1; }
-    if (ijk[2] >= this->nz_) { return -1; }
-
-    return ijk[0] + nx_*(ijk[1] + ny_*ijk[2]);
-}
-
-CellCollection::IJKTuple
-CellCollection::ind2sub(const std::size_t globalCell) const
-{
-    assert (globalCell < numGlobalCells());
-
-    auto ijk = IJKTuple{};
-    auto g   = globalCell;
-
-    ijk[0] = g % this->nx_;  g /= this->nx_;
-    ijk[1] = g % this->ny_;
-    ijk[2] = g / this->ny_;  assert (ijk[2] < this->nz_);
-
-    assert (globIdx(ijk) == globalCell);
-
-    return ijk;
-}
-
-// =====================================================================
 
 /// Implementation of ECLGraph interface.
 class Opm::ECLGraph::Impl
